Fix delay
This change fixes the issue where a server returns a decimal Retry-After, for example sub-second delays. The previous code would first round the delay to the nearest integer (integer type-cast) and THEN convert to microseconds. However to retain the decimal precision, we should first convert to microseconds and THEN cast to an integer.